+++ src/pages/home/sections/Callstack/Callstack.modal.tsx
@@ -1,5 +1,5 @@
 import * as Styled from './Callstack.styled.ts';
-import InfoModal from '../../../../components/Modal/Modal.tsx';
+import InfoModal from 'components/Modal/Modal.tsx';
 
 function CallStackModal({
 	isOpened,
