+++ src/pages/home/sections/MicroTasksQueue/MicroTasksQueue.styled.ts
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 import { css } from '@emotion/react';
-import InfoClosed from '../../../../components/CloseIcon/InfoIcon.tsx';
+import InfoClosed from 'components/CloseIcon/InfoIcon.tsx';
 
 export const MicroTasksQueue = styled.div`
 	flex: 1;
