+++ src/pages/home/sections/RequestAnimationFrameQueue/RequestAnimationFrameQueue.tsx
@@ -1,6 +1,6 @@
 import * as Styled from './RequestAnimationFrameQueue.styled.ts';
-import InfoIcon from '../../../../components/InfoIcon/InfoIcon.tsx';
-import useBoolean from '../../../../hooks/useBoolean.tsx';
+import InfoIcon from 'components/InfoIcon/InfoIcon.tsx';
+import useBoolean from 'utils/hooks/useBoolean.tsx';
 import { Zoom } from '@mui/material';
 import { List } from '../../Home.styled.ts';
 import RequestAnimationFrameQueueModal from './RequestAnimationFrameQueue.modal.tsx';
