+++ src/pages/home/sections/WebApiQueue/WebApiQueue.tsx
@@ -1,7 +1,7 @@
 import WebApiTask from './WebApiTask.tsx';
 import * as Styled from './WebApiQueue.styled.ts';
-import InfoIcon from '../../../../components/InfoIcon/InfoIcon.tsx';
-import useBoolean from '../../../../hooks/useBoolean.tsx';
+import InfoIcon from 'components/InfoIcon/InfoIcon.tsx';
+import useBoolean from 'utils/hooks/useBoolean.tsx';
 import { Zoom } from '@mui/material';
 import { List } from '../../Home.styled.ts';
 import WebApiQueueModal from './WebApiQueue.modal.tsx';
